Align colons & bookmarks when displaying, even when id's are not single-digits
Before this change, when Clink displayed bookmarks, the alignment between title
and url would offset by one or more characters if the id was larger than a
one-digit number. To fix this, I modified the listLinks function to
'right-align' the ids when displaying, so that all the colons lined up, and to
pad the urls to the right exactly as needed to line them up with their titles.
Speaking of alignment, I also placed all source comments on column 47...